requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.0.1
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
